requirements.txt
alembic==1.11.1
amqp==5.1.1
async-timeout==4.0.3
autocomplete==0.0.104
Babel==2.12.1
billiard==4.1.0
blinker==1.6.2
bottle==0.12.25
cachelib==0.10.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.4.1
dominate==2.8.0
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Moment==1.0.5
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
mysql-connector-python==8.1.0
oauthlib==2.1.0
packaging==23.1
pi==0.1.2
prompt-toolkit==3.0.39
protobuf==4.21.12
psycopg2-binary==2.9.6
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==5.0.0
requests==2.31.0
requests-oauthlib==1.1.0
rq==1.15.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.3.6
WTForms==3.0.1
